Fixed incorrect animation
... after recreating an activity
