+++ bubbletabbar/src/main/java/io/ak1/Bubble.kt
@@ -1,20 +1,22 @@
 package io.ak1
 
+import android.animation.ValueAnimator
 import android.annotation.SuppressLint
 import android.content.Context
 import android.graphics.Color
+import android.graphics.Rect
 import android.util.Log
 import android.view.Gravity
 import android.view.View
+import android.view.animation.LinearInterpolator
 import android.widget.FrameLayout
 import android.widget.ImageView
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.core.content.res.ResourcesCompat
 import io.ak1.parser.MenuItem
-import io.ak1.util.collapse
-import io.ak1.util.expand
 import io.ak1.util.setColorStateListAnimator
+import io.ak1.util.setCustomBackground
 
 @SuppressLint("ViewConstructor")
 class Bubble(context: Context, private var item: MenuItem) : FrameLayout(context) {
@@ -27,6 +29,10 @@ class Bubble(context: Context, private var item: MenuItem) : FrameLayout(context
     private val dpAsPixelsVertical = item.verticalPadding.toInt()
     private val dpAsPixelsIcons = item.iconSize.toInt()
     private val dpAsIconPadding = item.iconPadding.toInt()
+
+    private var expandAnimator: ValueAnimator? = null
+    private val DURATION = 350L
+    private val ALPHA = 0.15f
 
     init {
         id = item.id
@@ -103,10 +109,66 @@ class Bubble(context: Context, private var item: MenuItem) : FrameLayout(context
 
     override fun setSelected(selected: Boolean) {
         super.setSelected(selected)
+        this.clearAnimation()
+        expandAnimator?.removeAllUpdateListeners()
+        expandAnimator?.cancel()
         if (selected) {
-            title.expand(container, item.iconColor, item.cornerRadius)
+            expand(container, item.iconColor, item.cornerRadius)
         } else {
-            title.collapse(container, item.iconColor, item.cornerRadius)
+            collapse(container, item.iconColor, item.cornerRadius)
+        }
+    }
+
+    fun expand(container: LinearLayout, iconColor: Int, cornerRadius: Float) {
+        val bounds = Rect()
+        container.setCustomBackground(iconColor, ALPHA, cornerRadius)
+        title.apply {
+            paint.apply {
+                getTextBounds(text.toString(), 0, text.length, bounds)
+                expandAnimator = ValueAnimator.ofInt(0, bounds.width() + paddingStart + 10).apply {
+                    addUpdateListener {
+                        if (it.animatedFraction == (0.0f)) {
+                            visibility = View.INVISIBLE
+                        }
+                        layoutParams.apply {
+                            width = it.animatedValue as Int
+                        }
+
+                        if (it.animatedFraction == (1.0f)) {
+                            visibility = View.VISIBLE
+                        }
+                        requestLayout()
+                    }
+                    interpolator = LinearInterpolator()
+
+                    duration = DURATION
+                }
+                expandAnimator?.start()
+            }
+        }
+    }
+
+    fun collapse(container: LinearLayout, iconColor: Int, cornerRadius: Float) {
+        title.apply {
+            animate().alpha(0f).apply {
+                setUpdateListener {
+                    layoutParams.apply {
+                        width = (width - (width * it.animatedFraction)).toInt()
+                    }
+                    if (it.animatedFraction == 1.0f) {
+                        visibility = View.GONE
+                        alpha = 1.0f
+                    }
+                    interpolator = LinearInterpolator()
+                    duration = DURATION
+                    container.setCustomBackground(
+                        iconColor,
+                        ALPHA - (ALPHA * it.animatedFraction),
+                        cornerRadius
+                    )
+                    requestLayout()
+                }
+            }.start()
         }
     }
